test: StubSockect 으로 serialize 코드 옮김
